% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_neiprob.R
\name{calc_neiprob}
\alias{calc_neiprob}
\title{Calculating a set of neighbor QTL effects from conditional genotype probabilities}
\usage{
calc_neiprob(
  genoprobs,
  gmap,
  a2,
  d2,
  contrasts = c(TRUE, TRUE, TRUE),
  smap,
  scale,
  grouping = rep(1, nrow(smap))
)
}
\arguments{
\item{genoprobs}{Conditional genotype probabilities as obtained from \code{qtl2::calc_genoprob()}.}

\item{gmap}{Genetic map including observed and pseudomarkers, as obtained from \code{qtl2::insert_pseudomarkers()}.}

\item{a2}{A numeric scalar indicating additive deviation.}

\item{d2}{A numeric scalar indicating dominance deviation.}

\item{contrasts}{A vector composed of three TRUE/FALSE values, which represents the presence/absence of specific genotypes as c(TRUE/FALSE, TRUE/FALSE, TRUE/FALSE) = AA, AB, BB.}

\item{smap}{A matrix showing a spatial map. The first and second column include spatial points along a x-axis and y-axis, respectively.}

\item{scale}{A numeric scalar indicating the maximum spatial distance between a focal individual and neighbors to define neighbor effects.}

\item{grouping}{An integer vector assigning each individual to a group. This argument can be useful when \code{smap} contains different experimental replicates. Default setting means that all individuals are belong to a single group.}
}
\value{
A numeric matrix containing individuals x marker elements for neighbor QTL effects.
}
\description{
A function to calculate self QTL effects for all individuals, with given deviation coefficients and conditional genotype probabilities.
}
\examples{
set.seed(1234)
data("fake.f2",package="qtl")
fake_f2 <- qtl2::convert2cross2(fake.f2)
fake_f2 <- subset(fake_f2,chr=c(1:19))
smap_f2 <- cbind(runif(qtl2::n_ind(fake_f2),1,100),runif(qtl2::n_ind(fake_f2),1,100))
gmap_f2 <- qtl2::insert_pseudomarkers(fake_f2$gmap, step=2)
genoprobs_f2 <- qtl2::calc_genoprob(fake_f2,gmap_f2)

neiprobs <- calc_neiprob(genoprobs=genoprobs_f2,
                         gmap=gmap_f2, a2=1, d2=0,
                         smap=smap_f2, scale=20,
                         grouping=fake_f2$covar$pgm
                         )
}
\author{
Yasuhiro Sato (\email{sato.yasuhiro.36c@kyoto-u.jp})
}
