% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_pve.R
\name{calc_pve}
\alias{calc_pve}
\title{Calculating proportion of phenotypic variation explained (PVE) by neighbor effects}
\usage{
calc_pve(
  genoprobs,
  pheno,
  gmap,
  contrasts = c(TRUE, TRUE, TRUE),
  smap,
  s_seq,
  addcovar = NULL,
  grouping = rep(1, nrow(smap)),
  response = "quantitative",
  fig = TRUE
)
}
\arguments{
\item{genoprobs}{Conditional genotype probabilities as obtained from \code{qtl2::calc_genoprob()}.}

\item{pheno}{A vector of individual phenotypes.}

\item{gmap}{Genetic map including observed and pseudomarkers, as obtained from \code{qtl2::insert_pseudomarkers()}.}

\item{contrasts}{A vector composed of three TRUE/FALSE values. Depending on crossing design, it represents the presence/absence of specific genotypes as c(TRUE/FALSE, TRUE/FALSE, TRUE/FALSE) = AA, AB, BB.}

\item{smap}{A matrix showing a spatial map. The first and second column include spatial points along a x-axis and y-axis, respectively.}

\item{s_seq}{A numeric vector including a set of the maximum spatial distance between a focal individual and neighbors to define neighbor effects. A scalar is also allowed.}

\item{addcovar}{An optional matrix including additional non-genetic covariates. It contains no. of individuals x no. of covariates.}

\item{grouping}{An optional integer vector assigning each individual to a group. This argument can be useful when \code{smap} contains different experimental replicates. Default setting means that all individuals are belong to a single group.}

\item{response}{An optional argument to select trait types. The \code{"quantitative"} or \code{"binary"} applies the \code{"lmm.aireml()"} or \code{"logistic.mm.aireml()"} for a mixed model, respectively.}

\item{fig}{TRUE/FALSE to add a figure of PVE or not.}
}
\value{
A matrix containing the maximum neighbor distance, PVE by neighbor effects, and p-value by LRT.
\itemize{
\item{\code{scale}} {Maximum neighbor distance given as an argument}
\item{\code{PVE_nei}} {Proportion of phenotypic variation explained (PVE) by neighbor effects}
\item{\code{p-value}} {p-value by a likelihood ratio test between models with or without neighbor effects}
}
}
\description{
A function to calculate PVE by neighbor effects for a series of neighbor distance (\code{s_seq}) using a linear mixed model.
}
\details{
This function uses mixed models via the \code{gaston} package (Perdry & Dandine-Roulland 2020).
If \code{"binary"} is selected, \code{logistic.mm.aireml()} is called via the \code{gaston} package (see Chen et al. 2016 for the theory).
In such a case, \code{PVEnei} below is given by the variance component parameter \eqn{\sigma} (i.e., not a proportional value) and p-values are not provided.
}
\examples{
set.seed(1234)
data("fake.f2",package="qtl")
fake_f2 <- qtl2::convert2cross2(fake.f2)
fake_f2 <- subset(fake_f2,chr=c(1:19))
smap_f2 <- cbind(runif(qtl2::n_ind(fake_f2),1,100),runif(qtl2::n_ind(fake_f2),1,100))
gmap_f2 <- qtl2::insert_pseudomarkers(fake_f2$gmap, step=2)
genoprobs_f2 <- qtl2::calc_genoprob(fake_f2,gmap_f2)
s_seq <- quantile(dist(smap_f2),c(0.1*(1:10)))
pve_f2 <- calc_pve(genoprobs=genoprobs_f2,
                   pheno=fake_f2$pheno[,1], gmap=gmap_f2,
                   smap=smap_f2, s_seq=s_seq,
                   contrasts=c(TRUE,TRUE,TRUE),
                   addcovar=as.matrix(fake_f2$covar)
                   )
}
\references{
\itemize{
\item Perdry H, Dandine-Roulland C (2019) gaston: Genetic Data Handling (QC, GRM, LD, PCA) & Linear Mixed Models. R package version 1.5.5. https://CRAN.R-project.org/package=gaston
\item Chen H, Wang C, Conomos M. et al. (2016) Control for population structure and relatedness for binary traits in genetic association studies via logistic mixed models. The American Journal of Human Genetics 98: 653-666.
}
}
\author{
Yasuhiro Sato (\email{sato.yasuhiro.36c@kyoto-u.jp})
}
