% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_markers.R
\name{get_markers}
\alias{get_markers}
\title{Reshaping marker information}
\usage{
get_markers(gmap)
}
\arguments{
\item{gmap}{Genetic map including observed and pseudomarkers, as obtained from \code{qtl2::insert_pseudomarkers()}.}
}
\value{
A matrix showing the chromosome numbers (the first column) and positions (the second column) for all markers (row names).
}
\description{
A function to get marker information from a genetic map including observed and pseudo markers
}
\author{
Yasuhiro Sato (\email{sato.yasuhiro.36c@kyoto-u.jp})
}
