% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_neighbor.R
\name{int_neighbor}
\alias{int_neighbor}
\title{Testing marker-by-marker epistasis in neighbor QTL effects}
\usage{
int_neighbor(
  genoprobs,
  pheno,
  gmap,
  contrasts = c(TRUE, TRUE, TRUE),
  smap,
  scale,
  addcovar = NULL,
  addQTL,
  intQTL,
  grouping = rep(1, nrow(smap)),
  response = "quantitative"
)
}
\arguments{
\item{genoprobs}{Conditional genotype probabilities as taken from \code{qtl2::calc_genoprob()}.}

\item{pheno}{A vector of individual phenotypes.}

\item{gmap}{Genetic map including observed and pseudomarkers, as taken from \code{qtl2::insert_pseudomarkers()}.}

\item{contrasts}{A vector composed of three TRUE/FALSE values. Depending on crossing design, it represents the presence/absence of specific genotypes as c(TRUE/FALSE, TRUE/FALSE, TRUE/FALSE) = AA, AB, BB.}

\item{smap}{A matrix showing a spatial map. The first and second column include spatial points along a x-axis and y-axis, respectively.}

\item{scale}{A numeric scalar indicating the maximum spatial distance between a focal individual and neighbors to define neighbor effects.}

\item{addcovar}{An optional matrix including additional non-genetic covariates. It contains no. of individuals x no. of covariates.}

\item{addQTL}{A vector containing marker names that are considered covariates. This argument is necessary for \code{int_neighbor()}, and must match the marker names of \code{gmap}.}

\item{intQTL}{A name of a focal marker to be tested for its epistasis with the other markers in neighbor effects. The marker name must be included by \code{addQTL}.}

\item{grouping}{An optional integer vector assigning each individual to a group. This argument can be useful when \code{smap} contains different experimental replicates. Default setting means that all individuals are belong to a single group.}

\item{response}{An optional argument to select trait types. The \code{"quantitative"} or \code{"binary"} calls the \code{"gaussian"} or \code{"binomial"} family in \code{glm()}, respectively.}
}
\value{
A matrix of LOD scores for neighbor epistasis effects, with the chromosome numbers and positions. The row names correspond to marker names.
\itemize{
\item{\code{chr}} {Chromosome number}
\item{\code{pos}} {Marker position}
\item{\code{LOD_int}} {LOD score for epistasis in neighbor effects between a focal and the other markers}
}
}
\description{
A function to test interaction terms between one focal marker and the other markers across a genome.
}
\details{
This is an optinal function to test two-way interactions between the main neighbor effect of a focal marker given by \code{intQTL} and the others.
All the main neighbor effects are first esimated using \code{eff_neighbor()}, and then a two-way interaction term between the focal marker effect and its counterpart was considered an additional explnatory variable.
LOD score was compared between models with or without the two-way interaction.
}
\examples{
set.seed(1234)
data("fake.f2",package="qtl")
fake_f2 <- qtl2::convert2cross2(fake.f2)
fake_f2 <- subset(fake_f2,chr=c(1:19))
smap_f2 <- cbind(runif(qtl2::n_ind(fake_f2),1,100),runif(qtl2::n_ind(fake_f2),1,100))
gmap_f2 <- qtl2::insert_pseudomarkers(fake_f2$gmap, step=2)
genoprobs_f2 <- qtl2::calc_genoprob(fake_f2,gmap_f2)
int_f2 <- int_neighbor(genoprobs=genoprobs_f2,
                       pheno=fake_f2$pheno[,1],
                       gmap=gmap_f2, smap=smap_f2,
                       scale=19.37, addcovar=as.matrix(fake_f2$covar),
                       addQTL=c("D1M318","D1M212"), intQTL="D1M212"
                       )
plot_nei(int_f2, type="int")
}
\seealso{
scan_neighbor eff_neighbor
}
\author{
Yasuhiro Sato (\email{sato.yasuhiro.36c@kyoto-u.jp})
}
