% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik_glm.R
\name{logLik_glm.fit}
\alias{logLik_glm.fit}
\title{Calculating log-likelihood in generalized linear models}
\usage{
logLik_glm.fit(...)
}
\arguments{
\item{...}{Arguments to be passed to \code{glm.fit()}.}
}
\value{
Log-likelihood
}
\description{
An utility function to extract log-likelihood based on AIC of glm.fit()
}
