% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eff.R
\name{plot_eff}
\alias{plot_eff}
\title{Plot self and neighbor QTL effects across a genome}
\usage{
plot_eff(res, type = c("neighbor", "self"))
}
\arguments{
\item{res}{Output results of \code{eff_neighbor()}.}

\item{type}{An option to select \code{"self"} or \code{"neighbor"} effects to be shown. Default is \code{"neighbor"}.}
}
\description{
Plot estimated additive and dominance deviation for self or neighbor effects across a genome
}
\seealso{
eff_neighbor
}
\author{
Yasuhiro Sato (\email{sato.yasuhiro.36c@kyoto-u.jp})
}
