% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_nei.R
\name{plot_nei}
\alias{plot_nei}
\title{Plot a genome scan for self or neighbor QTL effect}
\usage{
plot_nei(res, type = "neighbor", chr = NULL, th = NULL, ...)
}
\arguments{
\item{res}{Output results of \code{scan_neighbor()}.}

\item{type}{Plot \code{"self"} or \code{"neighbor"} effects. Default is \code{"neighbor"} effects.}

\item{chr}{An optional vector to select chromosome numbers to be plotted. If \code{NULL}, shown are all chromosomes.}

\item{th}{Add genome-wide threshold by user-defined vectors or Bonferroni correction. Default is no thresholds added.}

\item{...}{Arguments to be passed to \code{plot()}.}
}
\description{
Plot LOD curves for a genome scan of self and neighbor QTL effects
}
\seealso{
scan_neighbor perm_neighbor
}
\author{
Yasuhiro Sato (\email{sato.yasuhiro.36c@kyoto-u.jp})
}
