% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selfprob.R
\name{selfprob}
\alias{selfprob}
\title{Calculating self QTL effects}
\usage{
selfprob(i, a1, d1, AA, AB, BB)
}
\arguments{
\item{i}{ID of a target individual.}

\item{a1}{A numeric scalar indicating additive deviation.}

\item{d1}{A numeric scalar indicating dominance deviation.}

\item{AA}{An individual x marker matrix of conditional probabilities for AA genotype.}

\item{AB}{An individual x marker matrix of conditional probabilities for AB genotype. Input NA if heterozygotes are absent.}

\item{BB}{An individual x marker matrix of conditional probabilities for BB genotype. Input NA for backcross lines.}
}
\value{
A numeric vector containing each marker effect for individual i.
}
\description{
A function to calculate self QTL effects for an individual, with given deviation coefficients and conditional genotype probabilities.
}
\author{
Yasuhiro Sato (\email{sato.yasuhiro.36c@kyoto-u.jp})
}
